Bullet Entity and renaming commonbullets
