+++ gtg/Bullet/pom.xml
@@ -7,15 +7,23 @@
     <artifactId>compiled-bundle-settings</artifactId>
     <version>1.0-SNAPSHOT</version>
   </parent>
+  <build>
+    <plugins>
+      <plugin>
+        <groupId>org.apache.maven.plugins</groupId>
+        <artifactId>maven-compiler-plugin</artifactId>
+      </plugin>
+    </plugins>
+  </build>
 
   <properties>
-    <bundle.symbolicName>dk.sdu.mmmi.swe.gtg.commonbullet</bundle.symbolicName>
-    <bundle.namespace>dk.sdu.mmmi.swe.gtg.commonbullet</bundle.namespace>
+    <bundle.symbolicName>dk.sdu.mmmi.swe.gtg.Bullet</bundle.symbolicName>
+    <bundle.namespace>dk.sdu.mmmi.swe.gtg.Bullet</bundle.namespace>
   </properties>
 
   <modelVersion>4.0.0</modelVersion>
   <groupId>dk.sdu.swe.gtg</groupId>
-  <artifactId>CommonBullet</artifactId>
+  <artifactId>Bullet</artifactId>
   <version>1.0-SNAPSHOT</version>
 
   <name>${bundle.symbolicName}</name>
@@ -36,7 +44,6 @@
       <groupId>org.osgi</groupId>
       <artifactId>osgi.core</artifactId>
       <optional>true</optional>
-      <scope>provided</scope>
     </dependency>
     <dependency>
       <groupId>dk.sdu.swe.gtg</groupId>
@@ -44,6 +51,8 @@
       <version>1.0-SNAPSHOT</version>
       <scope>compile</scope>
     </dependency>
+
+
   </dependencies>
 
 </project>
