+++ gtg/Bullet/src/main/java/dk/sdu/mmmi/swe/gtg/Bullet/Bullet.java
@@ -1,0 +1,5 @@
+package dk.sdu.mmmi.swe.gtg.Bullet;
+import dk.sdu.mmmi.swe.gtg.common.data.Entity;
+
+public class Bullet extends Entity {
+}
