+++ gtg/Bullet/src/main/java/dk/sdu/mmmi/swe/gtg/Bullet/BulletCreator.java
@@ -1,0 +1,4 @@
+package dk.sdu.mmmi.swe.gtg.Bullet;
+
+public class BulletCreator {
+}
