+++ gtg/Bullet/src/main/resources/readme.txt
@@ -1,0 +1,1 @@
+Bundle resources go here
