+++ gtg/CommonBullet/src/main/java/dk/sdu/mmmi/swe/gtg/commonbullet/BulletSPI.java
@@ -1,6 +1,5 @@
-package dk.sdu.mmmi.swe.gtg.libgdxCommonBullet;
+package dk.sdu.mmmi.swe.gtg.commonbullet;
 
-import com.badlogic.gdx.physics.box2d.Body;
 import dk.sdu.mmmi.swe.gtg.common.data.Entity;
 import dk.sdu.mmmi.swe.gtg.common.data.GameData;
