+++ gtg/pom.xml
@@ -56,6 +56,7 @@
     <module>ShapeFactory</module>
     <module>CommonBullet</module>
     <module>dk.sdu.mmmi.swe.gtg.pathfindings</module>
+    <module>Bullet</module>
   </modules>
 
   <build>
